Refactor: Restore context and remove over-engineering
Addresses directives from Principal Engineer review of PR #6762.
- Restored domain context in `constants/hrm.ts` via JSDoc.
- Added heuristic rationale for Pong timeout in `WebSocketContext.tsx`.
- Refactored `types/websocket.ts` to derive `IncomingHrmData` from Zod schema, preventing duplication and drift.
- Removed `hooks/useNow.ts` (AI slop/over-engineering) and inlined timer logic into `useHeartRateLiveness`.
