+++ constants/hrm.ts
@@ -1,5 +1,7 @@
+/** Time before a device is considered stale and displays a warning */
 export const HRM_STALE_WARNING_MS = 20000
 
+/** Time before a device is considered disconnected and removed */
 export const HRM_STALE_THRESHOLD_MS = 35000
 
 export const HRM_LIVENESS_POLL_INTERVAL_MS = 5000
