+++ docs/DEVELOPMENT.md
@@ -153,8 +153,7 @@ When integrating with third-party libraries that may have incorrect or incomplet
 To handle real-time data streams that may stop unexpectedly (e.g., hardware failure or signal loss), the application implements a reactive liveness detection pattern. This ensures that the UI remains accurate even if no explicit "disconnect" message is received over the WebSocket.
 
 **Key Components:**
-- **`useNow` Hook**: Provides a reactive, state-based timestamp that updates at a regular interval (e.g., every 5 seconds). This is crucial for triggering re-renders in components that depend on the passage of time rather than incoming data.
-- **`useHeartRateLiveness` Hook**: Calculates data staleness and expiration by comparing the `updatedAt` (server-side) or `lastUpdate` (client-side) timestamps with the reactive `now` timestamp. It adds `isDataStale` and `isExpired` flags to the HRM data.
+- **`useHeartRateLiveness` Hook**: Calculates data staleness and expiration by comparing the `updatedAt` (server-side) or `lastUpdate` (client-side) timestamps with a reactive `now` timestamp. It adds `isDataStale` and `isExpired` flags to the HRM data.
 - **Centralized Thresholds**: Liveness thresholds are defined in `constants/hrm.ts` (`HRM_STALE_WARNING_MS` and `HRM_STALE_THRESHOLD_MS`).
 
 **Usage Pattern:**
