+++ hooks/index.ts
@@ -7,6 +7,5 @@ export * from './useSpotifyRemoteExecution'
 export * from './useSpotifyWebPlayback'
 export * from './useVolumePreference'
 export * from './useWorkoutSession'
-export * from './useNow'
 export * from './useHeartRateLiveness'
 export * from './useFilteredHrmTiles'
